add model train and eval to data proc train and eval
